+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Feat
 
-- **agent**: allow passing identity into `pollForResponse`
+- **agent**: allow passing identity into all canister requests exposed by `HttpAgent`
 
 ## v5.2.0 (2026-03-24)